+++ apps/code-examples/src/app/code-examples/errors/error/embedded/demo.component.ts
@@ -8,7 +8,7 @@ import { SkyErrorModule } from '@skyux/errors';
   imports: [SkyErrorModule],
 })
 export class DemoComponent {
-  protected customAction(): void {
+  public customAction(): void {
     alert('action clicked!');
   }
 }